Give inherited-generic-members project test its own C# library
Two project tests previously shared `cs-test-assembly.csproj`. When
ghul-test runs them in parallel via `--use-dotnet-build`, both
`dotnet build` invocations hit MSBuild's `GenerateDepsFile` task
concurrently for the same csproj output, and the second fails with
'process cannot access the file...deps.json'. This race surfaced on
main after the test was added (PR gate happened to schedule them
serially).

Move the InheritedGenericMembers C# fixtures into a private
lib/lib.csproj inside the test directory, so the test no longer
shares a build target with anything else. cs-test-assembly stays
exclusive to ambiguous-method-hiding-1 (its original purpose).
