+++ project-tests/inherited-generic-members/inherited-generic-members.ghulproj
@@ -9,6 +9,6 @@
 
   <ItemGroup>
     <GhulSources Include="**/*.ghul" />
-    <ProjectReference Include="../cs-test-assembly/cs-test-assembly.csproj" />
+    <ProjectReference Include="lib/lib.csproj" />
   </ItemGroup>
 </Project>
